refactor(layout): split super-admin layout shell into focused components
